fix: controller fixed feat:add languages

+++ src/controllers/textract.js
@@ -1,48 +1,60 @@
-const { createWorker } = require('tesseract.js');
+const { createWorker } = require("tesseract.js");
 
 const imageRecognize = async (req, res) => {
-    try {
-        // Recognize text in image using Tesseract.js
-        const worker = await createWorker({
-            logger: m => console.log(m)
-        });
+  const { imageUrl, language } = req.body;
 
-        await worker.loadLanguage('eng');
-        await worker.initialize('eng');
-        const image = "https://tesseract.projectnaptha.com/img/eng_bw.png"
-        const { data: { text } } = await worker.recognize(image);
+  if (!imageUrl) {
+    return res.status(400).send("Missing imageUrl");
+  }
 
-        await worker.terminate();
+  if (!language) {
+    return res.status(400).send("Missing language");
+  }
 
-        // Send recognized text back to client
-        res.send(text);
-    } catch (error) {
-        res.status(500).send('An error occurred while processing the image.');
-    }
+  try {
+    // Recognize text in image using Tesseract.js
+    const worker = await createWorker({
+      logger: (m) => console.log(m),
+    });
+
+    await worker.loadLanguage(language);
+    await worker.initialize(language);
+
+    const myImage = imageUrl;
+
+    const {
+      data: { text },
+    } = await worker.recognize(myImage);
+
+    await worker.terminate();
+
+    res.status(200).send(text);
+  } catch (error) {
+    res.status(500).send("An error occurred while processing the image.");
+  }
 };
 
 const imageDetect = async (req, res) => {
-    try {
-        // Detect text in image using Tesseract.js
-        const path = require('path');
-        const Tesseract = require('tesseract.js');
+  try {
+    // Detect text in image using Tesseract.js
+    const path = require("path");
+    const Tesseract = require("tesseract.js");
 
-        const [, , imagePath] = process.argv;
-        // const image = path.resolve(__dirname, (imagePath || '../../tests/assets/images/cosmic.png'));
+    const [, , imagePath] = process.argv;
+    // const image = path.resolve(__dirname, (imagePath || '../../tests/assets/images/cosmic.png'));
 
-        const image = "https://tesseract.projectnaptha.com/img/eng_bw.png"
-        console.log(`Recognizing ${image}`);
+    const image = "https://tesseract.projectnaptha.com/img/eng_bw.png";
+    console.log(`Recognizing ${image}`);
 
-        Tesseract.detect(image, { logger: m => console.log(m) })
-            .then(({ data }) => {
-                console.log(data);
-            });
-    } catch (error) {
-        res.status(500).send('An error occurred while processing the image.');
-    }
+    Tesseract.detect(image, { logger: (m) => console.log(m) }).then(({ data }) => {
+      console.log(data);
+    });
+  } catch (error) {
+    res.status(500).send("An error occurred while processing the image.");
+  }
 };
 
-
 module.exports = {
-    imageRecognize, imageDetect
+  imageRecognize,
+  imageDetect,
 };
